# s3fifo add ghost fifo