--- a/pkg/fileservice/fifocache/data_cache.go
+++ b/pkg/fileservice/fifocache/data_cache.go
@@ -43,7 +43,17 @@ func NewDataCache(
 var seed = maphash.MakeSeed()
 
 func shardCacheKey(key fscache.CacheKey) uint64 {
-	hasher := new(maphash.Hash)
+
+	/*
+		hkey := []byte("12345678901234567890098765432112")
+		hasher, _ := highwayhash.New64(hkey)
+		hasher.Write(util.UnsafeToBytes(&key.Offset))
+		hasher.Write(util.UnsafeToBytes(&key.Sz))
+		hasher.Write([]byte(key.Path))
+		return hasher.Sum64()
+	*/
+
+	var hasher maphash.Hash
 	hasher.SetSeed(seed)
 	hasher.Write(util.UnsafeToBytes(&key.Offset))
 	hasher.Write(util.UnsafeToBytes(&key.Sz))