### **User description**
## What type of PR is this?

- [ ] API-change
- [ ] BUG
- [x] Improvement
- [ ] Documentation
- [ ] Feature
- [ ] Test and CI
- [ ] Code Refactoring

## Which issue(s) this PR fixes:

issue #21654 #21589

## What this PR does / why we need it:

1. add ghost fifo to s3fifo cache as stated in the paper.
2. clean up shard map to avoid race conditions with items under different locks (shards locks and queue lock).  
3. each item has its own mutex to protect postFunc() and its data and improve the concurrency.
4. all queues have their own mutex to improve concurrency.
5. replace VERY SLOW CompareAndSwap() with mutex to protect CacheItem for increment and decrement the frequency.
6. introduce disable-s3fifo in config.


___

### **PR Type**
Enhancement


___

### **Description**
- Introduced a new `ghost` structure for S3FIFO caching.

- Enhanced `Cache` to integrate `ghost` for better eviction handling.

- Added logic to manage `ghost` entries during enqueue and eviction.

- Implemented a new file `ghost.go` for the `ghost` structure.


___



### **Changes walkthrough** 📝
<table><thead><tr><th></th><th align="left">Relevant files</th></tr></thead><tbody><tr><td><strong>Enhancement</strong></td><td><table>
<tr>
  <td>
    <details>
      <summary><strong>fifo.go</strong><dd><code>Integrate `ghost` structure into S3FIFO cache</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </dd></summary>
<hr>

pkg/fileservice/fifocache/fifo.go

<li>Added a <code>ghost</code> field to the <code>Cache</code> structure.<br> <li> Integrated <code>ghost</code> logic into enqueue and eviction processes.<br> <li> Improved cache management by leveraging <code>ghost</code> for better eviction <br>handling.


</details>


  </td>
  <td><a href="https://github.com/matrixorigin/matrixone/pull/21653/files#diff-02fe777a115c8324898d3dfe7680a9227e5d16afa823d0fb0e69351031f81842">+20/-4</a>&nbsp; &nbsp; </td>

</tr>

<tr>
  <td>
    <details>
      <summary><strong>ghost.go</strong><dd><code>Implement `ghost` structure for S3FIFO caching</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </dd></summary>
<hr>

pkg/fileservice/fifocache/ghost.go

<li>Introduced a new <code>ghost</code> structure for S3FIFO caching.<br> <li> Implemented methods for adding, removing, and checking keys in <code>ghost</code>.<br> <li> Added logic to manage the size of the <code>ghost</code> list.


</details>


  </td>
  <td><a href="https://github.com/matrixorigin/matrixone/pull/21653/files#diff-d27a1055e2cd99529d0e3f3b19cce2ad17dc7e2a65960a5fd77856dd87f15c72">+54/-0</a>&nbsp; &nbsp; </td>

</tr>
</table></td></tr></tr></tbody></table>

___

> <details> <summary>  Need help?</summary><li>Type <code>/help how to ...</code> in the comments thread for any questions about Qodo Merge usage.</li><li>Check out the <a href="https://qodo-merge-docs.qodo.ai/usage-guide/">documentation</a> for more information.</li></details>