@reusee 2025-04-09
对比 main 分支和这个 pr 的性能：
···
 benchstat ~/mo/pkg/fileservice/fifocache/b1 b2
name                   old time/op  new time/op  delta
SequentialSet-10        144ns ± 2%   330ns ± 2%  +129.05%  (p=0.008 n=5+5)
ParallelSet-10         52.9ns ± 2%  94.0ns ± 2%   +77.70%  (p=0.008 n=5+5)
Get-10                 14.3ns ± 1%  20.0ns ± 4%   +39.49%  (p=0.008 n=5+5)
ParallelGet-10         11.3ns ±10%  12.6ns ±12%      ~     (p=0.056 n=5+5)
ParallelGetOrSet-10    35.1ns ± 4%  54.1ns ± 1%   +53.96%  (p=0.008 n=5+5)
ParallelEvict-10        332ns ± 0%   142ns ± 1%   -57.31%  (p=0.008 n=5+5)
ShardCacheKey-10       18.4ns ± 1%  20.8ns ± 0%   +12.88%  (p=0.008 n=5+5)
EnsureNBytesAndSet-10  79.1ns ± 3%   0.5ns ±35%   -99.43%  (p=0.008 n=5+5)
DataCacheSet-10         361ns ± 3%   628ns ± 1%   +73.77%  (p=0.008 n=5+5)
DataCacheGet-10        33.4ns ± 1%  65.6ns ± 0%   +96.53%  (p=0.008 n=5+5)
···

下降太多了。这个会极大影响整体性能。fifocache用在内存缓存，和metadata缓存，性能是最重要的因素。

其中 EnsureNBytesAndSet 的耗时降低很多，似乎有正确性问题。

建议将这个pr拆分成多个小pr，每次只做一件事，并且确保不引入任何的性能下降。